+++ CHANGES.md
@@ -10,6 +10,8 @@
 - Added in-memory cache for Ion asset endpoint responses to avoid repeated requests.
 - Add ScopeGuard utility to automatically execute function when exiting a scope.
 - Add glTF copyright information to the list of tileset credits.
+- Credits are now sorted based on the number of occurrences.
+- Added option to show credits on screen.
 - Expose the swizzle string in feature texture property views.
 
 ### v0.13.0 - 2022-03-01